backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
azure-cognitiveservices-speech==1.38.0
openai==1.51.2
sqlalchemy==2.0.36
alembic==1.16.2
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.1
websockets==15.0.1
sse-starlette==2.1.5

